Move skiplist test into a directory
