+++ crossbeam-skiplist/src/set.rs
@@ -376,18 +376,3 @@ where
         write!(f, "Range {{ ... }}")
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::SkipSet;
-
-    #[test]
-    fn smoke() {
-        let m = SkipSet::new();
-        m.insert(1);
-        m.insert(5);
-        m.insert(7);
-    }
-
-    // TODO(stjepang): Write more tests.
-}
